+++ packages/components/pie-textarea/src/defs.ts
@@ -4,7 +4,19 @@ export const sizes = ['small', 'medium', 'large'] as const;
 export const resizeModes = ['auto', 'manual', 'none'] as const;
 export const statusTypes = ['default', 'success', 'error'] as const;
 
+type AriaProps = {
+    /**
+     * Provides an accessible label for the textarea. Should be used when no visible label is associated with the textarea.
+     */
+    label?: string;
+}
+
 export interface TextareaProps {
+    /**
+     * ARIA attributes for the textarea.
+     */
+    aria?: AriaProps;
+
     /**
      * Same as the HTML disabled attribute - indicates whether the textarea is disabled.
      */
@@ -93,7 +105,7 @@ export interface TextareaProps {
 /**
  * The default values for the `TextareaProps` that are required (i.e. they have a fallback value in the component).
  */
-type DefaultProps = ComponentDefaultProps<TextareaProps, keyof Omit<TextareaProps, 'name' | 'autocomplete' | 'assistiveText' | 'defaultValue' | 'maxlength' | 'rows'>>;
+type DefaultProps = ComponentDefaultProps<TextareaProps, keyof Omit<TextareaProps, 'name' | 'autocomplete' | 'assistiveText' | 'defaultValue' | 'maxlength' | 'rows' | 'aria'>>;
 
 /**
  * Default values for optional properties that have default fallback values in the component.